# swagger/openapi.yaml
openapi: 3.0.0
info:
  title: FacePay API
  version: 1.0.0
  description: |
    ## FacePay - Biometric Payment Platform API
    
    FacePay combines advanced biometric authentication with secure payment processing to create a seamless, secure payment experience. Our API provides comprehensive endpoints for user management, payment processing, biometric authentication, and transaction handling.
    
    ### Key Features
    - 🔐 **Biometric Authentication**: WebAuthn-based face recognition and fingerprint authentication
    - 💳 **Multi-Payment Support**: Stripe, crypto, and traditional payment methods
    - 🛡️ **Enhanced Security**: Advanced threat detection and anomaly analysis
    - 📊 **Analytics Dashboard**: Comprehensive payment and security analytics
    - 🔄 **Real-time Processing**: Instant transaction processing and validation
    
    ### Authentication
    Most endpoints require authentication using JWT tokens. Include your token in the Authorization header:
    ```
    Authorization: Bearer YOUR_JWT_TOKEN
    ```
    
    ### Rate Limiting
    - **Standard endpoints**: 1000 requests per hour per user
    - **Authentication endpoints**: 10 requests per minute per IP
    - **Payment endpoints**: 100 requests per hour per user
    
    ### Error Handling
    All endpoints return standardized error responses with appropriate HTTP status codes and detailed error information.
    
  termsOfService: https://facepay.com/terms
  contact:
    name: FacePay API Support
...
servers:
...
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user with email and password
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              standard_login:
                summary: Standard login
                value:
                  email: john.doe@example.com
                  password: securepassword123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AuthResponse'
              examples:
                login_success:
                  summary: Successful login
                  value:
                    success: true
                    data:
                      user:
                        id: clp1abc123def456ghi789
                        email: john.doe@example.com
                        name: John Doe
                        createdAt: "2024-01-01T12:00:00Z"
                        updatedAt: "2024-01-01T12:00:00Z"
                      tokens:
                        accessToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                        refreshToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                        expiresIn: 3600
                      sessionId: sess_abc123def456
                    message: Login successful
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '400':
          $ref: '#/components/responses/ValidationError'
        '429':
          description: Too many login attempts
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /payments/analytics:
    get:
      tags:
        - Analytics
      summary: Get payment analytics
      description: Detailed payment method and processing analytics
      operationId: getPaymentAnalytics
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Payment analytics retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalVolume:
                    type: number
                  transactionCount:
                    type: integer
                  averageAmount:
                    type: number
                  paymentMethodBreakdown:
                    type: object
                  successRateByMethod:
                    type: object

  # Admin Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from the login endpoint. Include it in the Authorization header:
        ```
        Authorization: Bearer YOUR_JWT_TOKEN
        ```
    
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key for external integrations and webhooks
    
    BiometricAuth:
      type: oauth2
      description: Biometric authentication using WebAuthn
      flows:
        authorizationCode:
          authorizationUrl: /api/webauthn/authenticate/start
          tokenUrl: /api/webauthn/authenticate/complete
          scopes:
            biometric:read: Read biometric authentication data
            biometric:write: Register and manage biometric credentials
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          additionalProperties: true
        message:
          type: string
          example: Operation completed successfully
      required:
        - success
        - data
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: An error occurred
        code:
          type: string
          example: INTERNAL_ERROR
        details:
          type: object
          additionalProperties: true
      required:
        - success
        - error
    ValidationError:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: Validation failed
        code:
          type: string
          example: VALIDATION_ERROR
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: email
              message:
                type: string
                example: Email is required

    # User Schemas
    User:
      type: object
      properties:
        id:
          type: string
          description: Unique user identifier
          example: clp1abc123def456ghi789
        email:
          type: string
          format: email
          description: User email address
          example: john.doe@example.com
        name:
          type: string
          description: User full name
          example: John Doe
        createdAt:
          type: string
          format: date-time
          description: Account creation timestamp
          example: 2024-01-01T12:00:00Z
        updatedAt:
          type: string
          format: date-time
          description: Last profile update timestamp
          example: 2024-01-01T12:00:00Z
        lastLogin:
          type: string
          format: date-time
          description: Last login timestamp
          example: 2024-01-01T12:00:00Z
...
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          description: User email address
          example: john.doe@example.com
        password:
          type: string
          minLength: 8
          description: User password
          example: securepassword123
      required:
        - email
        - password
...
    AuthResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        tokens:
          type: object
          properties:
            accessToken:
              type: string
              description: JWT access token
              example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
            refreshToken:
              type: string
              description: JWT refresh token
              example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
            expiresIn:
              type: integer
              description: Token expiration time in seconds
              example: 3600
        sessionId:
          type: string
          description: Unique session identifier
          example: sess_abc123def456

    # Payment Schemas
...
  responses:
    UnauthorizedError:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: Authentication required
            code: UNAUTHORIZED
...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
          example:
            success: false
            error: Validation failed
            code: VALIDATION_ERROR
            details:
              - field: email
                message: Email is required
              - field: password
                message: Password must be at least 8 characters
    InternalServerError:
      description: Internal server error occurred
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: Internal server error
            code: INTERNAL_ERROR
...
security:
  - BearerAuth: []